feat: SVG exercise icons, Strava integration, human body muscle diagram
- 38 custom SVG exercise icons (white geometric line-art, tinted at runtime by muscle group color)
- Full Strava OAuth2 integration: auth flow, token refresh, workout upload
- Deep link handler (fitforge://strava-callback) for Strava OAuth callback
- Human body muscle diagram with path-based silhouette (front + back views)
  showing primary (full color) and secondary (50% opacity) muscle highlights
- Strava connect/disconnect section in profile screen
- Upload to Strava toggle in post-workout finish dialog
- Workout card muscle chip display (replaced overlapping circles)
- exercise.dart: iconAsset getter mapping 55 exercise IDs to 38 SVG files
- pubspec.yaml: added flutter_svg, http, app_links dependencies

+++ lib/presentation/screens/progress/progress_screen.dart
@@ -81,7 +81,7 @@ class _SummaryCards extends StatelessWidget {
     final now = DateTime.now();
     final weekSessions = history.where((s) => s.startTime.isAfter(now.subtract(const Duration(days: 7)))).length;
     final totalSessions = history.length;
-    final totalVolume = history.fold(0, (sum, s) => sum + s.totalVolume);
+    final totalVolume = history.fold(0, (sum, s) => sum + (s.totalVolume as int));
 
     return Row(
       children: [
@@ -140,7 +140,7 @@ class _VolumeChart extends StatelessWidget {
       final weekEnd = weekStart.add(const Duration(days: 7));
       final weekVolume = history
           .where((s) => s.startTime.isAfter(weekStart) && s.startTime.isBefore(weekEnd))
-          .fold(0, (sum, s) => sum + s.totalVolume);
+          .fold(0, (sum, s) => sum + (s.totalVolume as int));
       return FlSpot(i.toDouble(), (weekVolume / 1000).toDouble());
     });
